apps/api/pyproject.toml
[project]
name = "knowledge-chatbox-api"
version = "0.1.0"
description = "Local-first API service for Knowledge Chatbox"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "asgi-correlation-id>=4.3.4",
    "pydantic-settings>=2.1.0",
    "uvicorn>=0.29.0",
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.0",
    "chromadb>=1.1.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.2.0",
    "openai>=1.40.0",
    "python-multipart>=0.0.9",
    "pillow>=10.4.0",
    "argon2-cffi>=23.1.0",
    "PyJWT>=2.10.1",
    "httpx>=0.28.1",
    "structlog>=25.5.0",
    "anthropic>=0.86.0",
    "ollama>=0.6.1",
]

[project.scripts]
knowledge-chatbox-api = "knowledge_chatbox_api:main"

[dependency-groups]
dev = [
    "basedpyright>=1.31.0",
    "pytest>=9.0.2",
    "pytest-httpx>=0.36.0",
    "ruff>=0.15.6",
]

[build-system]
requires = ["uv_build>=0.10.11,<0.11.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.basedpyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
include = ["src"]
exclude = ["tests"]
reportMissingTypeStubs = false
